contributors/ttwf_tokyo/starters/HTML5/samples/4.5.5.html

<!DOCTYPE HTML>
<html>
<head>
 <title>The OL Element</title>
 <meta http-equiv="Content-Type" content="text/html;charset=UTF-8">
 <link rel="help" href="http://www.w3.org/TR/html5/single-page.html#semantics">
</head>
<body>
<h1>The OL Element</h1>
<ol id='olTestTag' style='overflow: hidden'></ol>
<div id="log"></div>
<script src="http://www.w3c-test.org/resources/testharness.js"></script>
<script src="http://www.w3c-test.org/resources/testharnessreport.js"></script>
<script>
function reversedAttribute(){
    assert_equals("boolean", typeof document.getElementById("olTestTag").reversed);
}
function startAttribute(){
    assert_equals("number", typeof document.getElementById("olTestTag").start);
}
function typeAttribute(){
    assert_equals("string", typeof document.getElementById("olTestTag").type);
}

test(reversedAttribute, "HTMLOListElement reversed attribute");
test(startAttribute, "HTMLOListElement start attribute");
test(typeAttribute, "HTMLOListElement test attribute");
</script>
